moving anonymous objects to appstate
